fix logical separate component
